requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
umap-learn